support images in creations on frontend
Had to finally add some CSS for it to not break.

I make use of `instanse.trusted` list to decide whether to show the image or not.

I also added a "load more" button to not make too much requests to other instances - since each image is most likely a separate one.